Fully implement @import; move some functions to the right file
